chore: added folders to organize home and introduction view components
