+++ src/components/Home/components/TourInstructionsHome.tsx
@@ -1,21 +1,18 @@
-import { RefObject } from "react";
 import {
-  AboutParagraph,
-  AboutAnchorHeader,
+  HomeSubHeader,
+  HomeParagraph,
   StaticcontentButtonButton,
   StaticheaderBackgroundButton,
   StatictourButtonButton,
-} from "./styled_components";
+} from "../../styled_components";
 
-function TourInstructionsIntro(props: {
-  howToRef: RefObject<HTMLHeadingElement>;
-}) {
+function TourInstructionsHome() {
   return (
-    <section>
-      <AboutAnchorHeader id="how-to-take-the-tour" ref={props.howToRef}>
+    <>
+      <HomeSubHeader id="how-to-take-the-tour">
         How to Take the Tour
-      </AboutAnchorHeader>
-      <AboutParagraph>
+      </HomeSubHeader>
+      <HomeParagraph>
         If you choose to take the tour, please allow yourself{" "}
         <strong>between 2.5-3 hours</strong> to complete the walking tour all in
         one visit. You can start the tour by pressing the{" "}
@@ -23,8 +20,8 @@ function TourInstructionsIntro(props: {
         You may choose to only do a portion of the tour and come back at a later
         time. The tour will remember which markers you have already visited and
         allow you to begin where you left off.
-      </AboutParagraph>
-      <AboutParagraph>
+      </HomeParagraph>
+      <HomeParagraph>
         The first stop on the tour is at <strong>first marker location</strong>.
         From there the map will prompt you to the next stop with an orange
         marker. You may choose to visit any marker at any time but the loop we
@@ -35,15 +32,15 @@ function TourInstructionsIntro(props: {
           Mark as completed
         </StaticheaderBackgroundButton>{" "}
         check box on the content page.
-      </AboutParagraph>
-      <AboutParagraph>
+      </HomeParagraph>
+      <HomeParagraph>
         Not able to walk the tour? All materials are available at any time,
         simply click the{" "}
         <StatictourButtonButton>Content</StatictourButtonButton> button to watch
         at your own pace.
-      </AboutParagraph>
-    </section>
+      </HomeParagraph>
+    </>
   );
 }
 
-export default TourInstructionsIntro;
+export default TourInstructionsHome;
