+++ src/components/Introduction/Introduction.tsx
@@ -1,14 +1,17 @@
 import { useLocation } from "wouter";
 import { useRef } from "react";
-import { MainContainer, BackButton } from "./styled_components";
-import Header from "./Header";
-import Footer from "./Footer";
-import TourInstructionsIntro from "./TourInstructionsIntro";
-import IntroNav from "./IntroNav";
-import { intro } from "../services";
-import About from "./About";
-import Statement from "./Statement";
-import References from "./References";
+
+import Header from "../Header";
+import Footer from "../Footer";
+import TourInstructionsIntro from "./components/TourInstructionsIntro";
+import IntroNav from "./components/IntroNav";
+import About from "./components/About";
+import Statement from "./components/Statement";
+import References from "./components/References";
+
+import { MainContainer, BackButton } from "../styled_components";
+
+import { intro } from "../../services";
 
 function Introduction() {
   const [, setLocation] = useLocation();
@@ -38,13 +41,9 @@ function Introduction() {
             statementRef={statementRef}
             referencesRef={referencesRef}
           />
-
           <TourInstructionsIntro howToRef={howToRef} />
-
           <About aboutRef={aboutRef} />
-
           <Statement statementRef={statementRef} />
-
           <References referencesRef={referencesRef} />
         </article>
       </MainContainer>
