+++ src/components/Introduction/components/About.tsx
@@ -1,5 +1,5 @@
 import { RefObject } from "react";
-import { AboutParagraph, AboutAnchorHeader } from "./styled_components";
+import { AboutParagraph, AboutAnchorHeader } from "../../styled_components";
 
 function About(props: { aboutRef: RefObject<HTMLHeadingElement> }) {
   return (
