+++ src/components/Introduction/components/TourInstructionsIntro.tsx
@@ -1,18 +1,21 @@
+import { RefObject } from "react";
 import {
-  HomeSubHeader,
-  HomeParagraph,
+  AboutParagraph,
+  AboutAnchorHeader,
   StaticcontentButtonButton,
   StaticheaderBackgroundButton,
   StatictourButtonButton,
-} from "./styled_components";
+} from "../../styled_components";
 
-function TourInstructionsHome() {
+function TourInstructionsIntro(props: {
+  howToRef: RefObject<HTMLHeadingElement>;
+}) {
   return (
-    <>
-      <HomeSubHeader id="how-to-take-the-tour">
+    <section>
+      <AboutAnchorHeader id="how-to-take-the-tour" ref={props.howToRef}>
         How to Take the Tour
-      </HomeSubHeader>
-      <HomeParagraph>
+      </AboutAnchorHeader>
+      <AboutParagraph>
         If you choose to take the tour, please allow yourself{" "}
         <strong>between 2.5-3 hours</strong> to complete the walking tour all in
         one visit. You can start the tour by pressing the{" "}
@@ -20,8 +23,8 @@ function TourInstructionsHome() {
         You may choose to only do a portion of the tour and come back at a later
         time. The tour will remember which markers you have already visited and
         allow you to begin where you left off.
-      </HomeParagraph>
-      <HomeParagraph>
+      </AboutParagraph>
+      <AboutParagraph>
         The first stop on the tour is at <strong>first marker location</strong>.
         From there the map will prompt you to the next stop with an orange
         marker. You may choose to visit any marker at any time but the loop we
@@ -32,15 +35,15 @@ function TourInstructionsHome() {
           Mark as completed
         </StaticheaderBackgroundButton>{" "}
         check box on the content page.
-      </HomeParagraph>
-      <HomeParagraph>
+      </AboutParagraph>
+      <AboutParagraph>
         Not able to walk the tour? All materials are available at any time,
         simply click the{" "}
         <StatictourButtonButton>Content</StatictourButtonButton> button to watch
         at your own pace.
-      </HomeParagraph>
-    </>
+      </AboutParagraph>
+    </section>
   );
 }
 
-export default TourInstructionsHome;
+export default TourInstructionsIntro;
